Tighten local patch DSL support
